move repeat code block to common file
